+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Generate token from token
 + Add Refreh Token route
 + Get expiry date from token
++ Add Expiry at on jwt response on login - refresh
 
 
 ## 0.11.0